Reuse onboarding event checklist in settings
